+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.0.3-rc (TBA)
 ===
 * `entity/content_type`: PredefinedValues validation error message
+* `entity/content_type`: Range validation greater/less than equal to support
 
 v0.0.2 (2017-03-21)
 ===